int and real
