+++ z3/src/ast/mod.rs
@@ -13,23 +13,13 @@ use z3_sys::*;
 
 use crate::{Context, FuncDecl, IsNotApp, Pattern, Sort, SortDiffers, Symbol};
 
-use num::{bigint::BigInt, rational::BigRational};
-
 mod bool;
+mod int;
+mod real;
 
 pub use bool::Bool;
-
-/// [`Ast`] node representing an integer value.
-pub struct Int {
-    pub(crate) ctx: Context,
-    pub(crate) z3_ast: Z3_ast,
-}
-
-/// [`Ast`] node representing a real value.
-pub struct Real {
-    pub(crate) ctx: Context,
-    pub(crate) z3_ast: Z3_ast,
-}
+pub use int::Int;
+pub use real::Real;
 
 /// [`Ast`] node representing a float value.
 pub struct Float {
@@ -556,48 +546,6 @@ impl_ast!(Seq);
 impl_from_try_into_dynamic!(Seq, as_seq);
 impl_ast!(Regexp);
 
-impl Int {
-    pub fn from_big_int(ctx: &Context, value: &BigInt) -> Int {
-        Int::from_str(ctx, &value.to_str_radix(10)).unwrap()
-    }
-
-    pub fn from_str(ctx: &Context, value: &str) -> Option<Int> {
-        let sort = Sort::int(ctx);
-        let ast = unsafe {
-            let int_cstring = CString::new(value).unwrap();
-            let numeral_ptr = Z3_mk_numeral(ctx.z3_ctx.0, int_cstring.as_ptr(), sort.z3_sort);
-            if numeral_ptr.is_null() {
-                return None;
-            }
-
-            numeral_ptr
-        };
-        Some(unsafe { Int::wrap(ctx, ast) })
-    }
-}
-
-impl Real {
-    pub fn from_big_rational(ctx: &Context, value: &BigRational) -> Real {
-        let num = value.numer();
-        let den = value.denom();
-        Real::from_real_str(ctx, &num.to_str_radix(10), &den.to_str_radix(10)).unwrap()
-    }
-
-    pub fn from_real_str(ctx: &Context, num: &str, den: &str) -> Option<Real> {
-        let sort = Sort::real(ctx);
-        let ast = unsafe {
-            let fraction_cstring = CString::new(format!("{num:} / {den:}")).unwrap();
-            let numeral_ptr = Z3_mk_numeral(ctx.z3_ctx.0, fraction_cstring.as_ptr(), sort.z3_sort);
-            if numeral_ptr.is_null() {
-                return None;
-            }
-
-            numeral_ptr
-        };
-        Some(unsafe { Real::wrap(ctx, ast) })
-    }
-}
-
 impl Float {
     // Create a 32-bit (IEEE-754) Float [`Ast`] from a rust f32
     pub fn from_f32(ctx: &Context, value: f32) -> Float {
@@ -664,233 +612,6 @@ fn _atleast(ctx: &Context, args: &[Z3_ast], k: u32) -> Bool {
                 k,
             ),
         )
-    }
-}
-
-impl Int {
-    pub fn new_const<S: Into<Symbol>>(ctx: &Context, name: S) -> Int {
-        let sort = Sort::int(ctx);
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_const(ctx.z3_ctx.0, name.into().as_z3_symbol(ctx), sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn fresh_const(ctx: &Context, prefix: &str) -> Int {
-        let sort = Sort::int(ctx);
-        unsafe {
-            Self::wrap(ctx, {
-                let pp = CString::new(prefix).unwrap();
-                let p = pp.as_ptr();
-                Z3_mk_fresh_const(ctx.z3_ctx.0, p, sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn from_i64(ctx: &Context, i: i64) -> Int {
-        let sort = Sort::int(ctx);
-        unsafe { Self::wrap(ctx, Z3_mk_int64(ctx.z3_ctx.0, i, sort.z3_sort)) }
-    }
-
-    pub fn from_u64(ctx: &Context, u: u64) -> Int {
-        let sort = Sort::int(ctx);
-        unsafe { Self::wrap(ctx, Z3_mk_unsigned_int64(ctx.z3_ctx.0, u, sort.z3_sort)) }
-    }
-
-    pub fn as_i64(&self) -> Option<i64> {
-        unsafe {
-            let mut tmp: ::std::os::raw::c_longlong = 0;
-            if Z3_get_numeral_int64(self.ctx.z3_ctx.0, self.z3_ast, &mut tmp) {
-                Some(tmp)
-            } else {
-                None
-            }
-        }
-    }
-
-    pub fn as_u64(&self) -> Option<u64> {
-        unsafe {
-            let mut tmp: ::std::os::raw::c_ulonglong = 0;
-            if Z3_get_numeral_uint64(self.ctx.z3_ctx.0, self.z3_ast, &mut tmp) {
-                Some(tmp)
-            } else {
-                None
-            }
-        }
-    }
-
-    pub fn from_real(ast: &Real) -> Int {
-        unsafe { Self::wrap(&ast.ctx, Z3_mk_real2int(ast.ctx.z3_ctx.0, ast.z3_ast)) }
-    }
-
-    /// Create a real from an integer.
-    /// This is just a convenience wrapper around
-    /// [`Real::from_int()`]; see notes there.
-    pub fn to_real(&self) -> Real {
-        Real::from_int(self)
-    }
-
-    /// Create an integer from a bitvector.
-    ///
-    /// Signed and unsigned version.
-    ///
-    /// # Examples
-    /// ```
-    /// # use z3::{ast, Config, Context, SatResult, Solver};
-    /// # use z3::ast::Ast;
-    /// # let cfg = Config::new();
-    /// # let ctx = Context::new(&cfg);
-    /// # let solver = Solver::new(&ctx);
-    /// let bv = ast::BV::new_const(&ctx, "x", 32);
-    /// solver.assert(&bv._eq(&ast::BV::from_i64(&ctx, -3, 32)));
-    ///
-    /// let x = ast::Int::from_bv(&bv, true);
-    ///
-    /// assert_eq!(solver.check(), SatResult::Sat);
-    /// let model = solver.get_model().unwrap();
-    ///
-    /// assert_eq!(-3, model.eval(&x, true).unwrap().as_i64().unwrap());
-    /// ```
-    pub fn from_bv(ast: &BV, signed: bool) -> Int {
-        unsafe {
-            Self::wrap(&ast.ctx, {
-                Z3_mk_bv2int(ast.ctx.z3_ctx.0, ast.z3_ast, signed)
-            })
-        }
-    }
-
-    /// Create a bitvector from an integer.
-    /// This is just a convenience wrapper around
-    /// [`BV::from_int()`]; see notes there.
-    pub fn to_ast(&self, sz: u32) -> BV {
-        BV::from_int(self, sz)
-    }
-
-    varop! {
-        add(Z3_mk_add, Self);
-        sub(Z3_mk_sub, Self);
-        mul(Z3_mk_mul, Self);
-    }
-    unop! {
-        unary_minus(Z3_mk_unary_minus, Self);
-    }
-    binop! {
-        div(Z3_mk_div, Self);
-        rem(Z3_mk_rem, Self);
-        modulo(Z3_mk_mod, Self);
-        power(Z3_mk_power, Real);
-        lt(Z3_mk_lt, Bool);
-        le(Z3_mk_le, Bool);
-        gt(Z3_mk_gt, Bool);
-        ge(Z3_mk_ge, Bool);
-    }
-    // Z3 does support mixing ints and reals in add(), sub(), mul(), div(), and power()
-    //   (but not rem(), modulo(), lt(), le(), gt(), or ge()).
-    // TODO: we could consider expressing this by having a Numeric trait with these methods.
-    //    Int and Real would have the Numeric trait, but not the other Asts.
-    // For example:
-    //   fn add(&self, other: &impl Numeric) -> Dynamic { ... }
-    // Note the return type would have to be Dynamic I think (?), as the exact result type
-    //   depends on the particular types of the inputs.
-    // Alternately, we could just have
-    //   Int::add_real(&self, other: &Real) -> Real
-    // and
-    //   Real::add_int(&self, other: &Int) -> Real
-    // This might be cleaner because we know exactly what the output type will be for these methods.
-}
-
-impl Real {
-    pub fn new_const<S: Into<Symbol>>(ctx: &Context, name: S) -> Real {
-        let sort = Sort::real(ctx);
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_const(ctx.z3_ctx.0, name.into().as_z3_symbol(ctx), sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn fresh_const(ctx: &Context, prefix: &str) -> Real {
-        let sort = Sort::real(ctx);
-        unsafe {
-            Self::wrap(ctx, {
-                let pp = CString::new(prefix).unwrap();
-                let p = pp.as_ptr();
-                Z3_mk_fresh_const(ctx.z3_ctx.0, p, sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn from_real(ctx: &Context, num: i32, den: i32) -> Real {
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_real(
-                    ctx.z3_ctx.0,
-                    num as ::std::os::raw::c_int,
-                    den as ::std::os::raw::c_int,
-                )
-            })
-        }
-    }
-
-    pub fn as_real(&self) -> Option<(i64, i64)> {
-        unsafe {
-            let mut num: i64 = 0;
-            let mut den: i64 = 0;
-            if Z3_get_numeral_small(self.ctx.z3_ctx.0, self.z3_ast, &mut num, &mut den) {
-                Some((num, den))
-            } else {
-                None
-            }
-        }
-    }
-
-    pub fn approx(&self, precision: usize) -> ::std::string::String {
-        let s = unsafe {
-            CStr::from_ptr(Z3_get_numeral_decimal_string(
-                self.ctx.z3_ctx.0,
-                self.z3_ast,
-                precision as _,
-            ))
-        }
-        .to_str()
-        .unwrap();
-        s.strip_suffix('?').unwrap_or(s).to_owned()
-    }
-    pub fn approx_f64(&self) -> f64 {
-        self.approx(17).parse().unwrap() // 17 decimal digits needed to get full f64 precision
-    }
-
-    pub fn from_int(ast: &Int) -> Real {
-        unsafe { Self::wrap(&ast.ctx, Z3_mk_int2real(ast.ctx.z3_ctx.0, ast.z3_ast)) }
-    }
-
-    /// Create an integer from a real.
-    /// This is just a convenience wrapper around
-    /// [`Int::from_real()`]; see notes there.
-    pub fn to_int(&self) -> Int {
-        Int::from_real(self)
-    }
-
-    unop! {
-        is_int(Z3_mk_is_int, Bool);
-    }
-
-    varop! {
-        add(Z3_mk_add, Self);
-        sub(Z3_mk_sub, Self);
-        mul(Z3_mk_mul, Self);
-    }
-    unop! {
-        unary_minus(Z3_mk_unary_minus, Self);
-    }
-    binop! {
-        div(Z3_mk_div, Self);
-        power(Z3_mk_power, Self);
-        lt(Z3_mk_lt, Bool);
-        le(Z3_mk_le, Bool);
-        gt(Z3_mk_gt, Bool);
-        ge(Z3_mk_ge, Bool);
     }
 }
